make PPQConnectPage folder and refactor interfaces
