+++ src/frontend/src/pages/PPQConnectPage/interfaces.ts
@@ -1,0 +1,5 @@
+import { IPPQResult } from '../../types/interfaces/ppq-result.interface';
+
+export interface IPagePropState {
+  result: IPPQResult;
+}
